requirements.txt
Flask==<version>
Flask-JWT-Extended==<version>
Flask-SQLAlchemy==<version>
python-dotenv==<version>

